cvobjectdetect: Draw the id for each object
Sometimes you may see that for a given object it has different ids. That
is totally expected and is the natural behavior of an object detector,
since objects detector don't use to respect the order of ids.

That described issue may be solved with an assignment algorithm like the
Hungarian method or with an object tracker (distinct from an object
detector)

Closes #16