cvscale: Add filter to create a resized copy of the original frame
For this implementation, a class GstCVDupScaleMeta has been added. The idea
with this new class is that elements can write scaled frames into a
GstCVDupScaleMeta to a buffer, and elements that support GstCVDupScaleMeta
can read the scaled frames and work on it.

One example of it is cvscale, which adds a scaled frame to the meta of the
buffer and gstcvobjectdetect can now read if a GstCVDupScaleMeta is present
and works on the scaled copy of the frame instead of working on the original
frame.

Scaling down frames to gain speed is a typical strategy in computer vision
and image processing. But other techniques like getting the pyramid* of an
image is also used. So a later work could be to implement a filter that
gets the pyramid of the input frame and writes it in a GstCVDupScaleMeta.

 (*) https://en.wikipedia.org/wiki/Pyramid_(image_processing)

This way, just swapping the cvscale for that new filter should just be
compatible with cvobjectdetect and other elements supporting
GstCVDupScaleMeta.

Pipeline example:

  gst-launch-1.0 v4l2src ! videoconvert ! cvscale scale-factor=0.7 ! \
      cvdlibobjectdetecthog model-location=m.dat ! videoconvert ! \
      xvimagesink