cvdraw: Add element to draw shapes contained in GstCVObjectInfoMap
A future work pretends to implement a new element implementing the
Hungarian method on the info output by gstobjectdetect, so the
order of the indexes/ids can be kept.

As you can imagine, that would mean duplicating code. Instead of
that, and to also, in some way, to follow the UNIX philosophy of
"doing one thing and doing it well", a new cvdraw element is added
that reads the GstCVObjectInfoMap output by cvobjectdetect and
takes this data to draw into the current frame.

This is a dummy element and should be only used for quick testing
purposes or visualization.

Closes #16