hooks: Add pre-commit hook to check coding style
This commit will be the base for future submission of code written
in C or C++ since they have to follow a coding style.

The future plan is to copy automatically the pre-commit.hook from
the meson file to .git/hooks/pre-commit. So this way we (almost)
force to developers to always submit code following our coding
style.

We use the same coding style used by GStreamer in C files and a
similar coding style for C++ files.

The tool gstcv-indent can be used as:

  ./scripts/gstcv-indent file.c
  ./scripts/gstcv-indent file.cpp

and it will automatically write changes in place.

Note: The used uncrustify config file is a courtesy of
Matthew Waters (@ystreet00)

Closes #1