cvobjectdetect: Put coordinates of detected faces in a buffer meta
Instead of just adding a new meta class containing a graphene_rect_t,
this commit tries to forecast the future and add a more complex
meta structure called GstCVObjectInfoMapMeta which contains just a
GstCVObjectInfoMap.

A GstCVObjectInfoMapMeta is a dictionary of indices (integers)
pointing to a list of GstCVObjectInfo. Each index is supposed to
identify each object, while each GstCVObjectInfo should contain
info about the detected object, which is not limited to only the
ROI (region-of-interest) but may be also the landmarks (in the case
a new element to detect the landmarks*, keypoints marking particular
parts of an object, gets added.

Also, each GstCVObjectInfo has a tag and knows about the GstElement
that created it. So, we can know, for example, if a ROI comes for
an object detect element or from a tracking element (when that gets
implemented).

 (*) https://en.wikipedia.org/wiki/Landmark_point