hooks: pre-commit: Prevent to parse .cpp files as .c files
In the pre-commit hook, a temporary file of the current staged
revision is created. Because the resulting filename is a "random"
string without an extension, uncrustify may think that the file is
written in other programming language instead of the one it is
actually written in. To prevent that, the file extension is added
to this temporary name.