cvobjectcache: Add element to cache object info maps per frame number
This element creates a new global cache (GstCVObjectInfoMapNamedCache).

When this element is created, a cache-id needs to be set, so a new entry
(a "cache") is created in the named cache (or get if already exists). In
this "cache", a copy of an object info map (GstCVObjectInfoMap), coming
from the buffer meta, will be inserted (filtered by a sub-key if
specified) mapped by frame number.

This process is repeated for each frame. The by-frame-number cache has a
maximum size, and insertion is done using FIFO.